+++ src/collector/mod.rs
@@ -100,11 +100,9 @@ mod top_score_collector;
 pub use self::top_collector::ComparableDoc;
 pub use self::top_score_collector::{TopDocs, TopNComputer};
 
-mod custom_score_top_collector;
-pub use self::custom_score_top_collector::{CustomScorer, CustomSegmentScorer};
-
+mod sort_key;
 mod sort_key_top_collector;
-pub use self::sort_key_top_collector::{SegmentSortKeyComputer, SortKeyComputer};
+pub use self::sort_key::{SegmentSortKeyComputer, SortKeyComputer};
 mod facet_collector;
 pub use self::facet_collector::{FacetCollector, FacetCounts};
 use crate::query::Weight;
